% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featureFinder.R
\name{featureFinder}
\alias{featureFinder}
\title{FeatureFinder}
\usage{
featureFinder(dataStorage, Structure, SeriesInstanceUID)
}
\description{
This class compares many mmButo data to find possibles features for predictors
}

