% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoLet.R
\name{GLT.getAttribute}
\alias{GLT.getAttribute}
\title{GLT.getAttribute - a wrapper function to get an attribute from a \code{geoLet} object}
\usage{
GLT.getAttribute(obj = obj, attribute = attribute,
  seriesInstanceUID = seriesInstanceUID, fileName = fileName)
}
\arguments{
\item{obj}{an \code{geoLet} object}

\item{attribute}{the name of the attribute to be retrieved. The available attributes arE:
\itemize{
  \item \code{dataStorage} is probably the most informative attribute. It stores all the information about images, ROIs, etc, archieved as a list of lists deeply nested
  \item \code{PatientName} is the Patient name as stored in the DICOM study
  \item \code{PatientID} is the Patient ID as stored in the DICOM study
  \item \code{PatientSex} is the sex of the patient
  \item \code{Rows} is the number of rows in the DICOM images
  \item \code{Columns} is the number of columns in the DICOM images
  \item \code{PixelSpacing} is xy dimension for each voxel
  \item \code{SliceThickness} is the thickness of the interested slice
  \item \code{ImagePositionPatient} is the ImagePositionPatient of the interested slice
  \item \code{ImageOrientationPatient} is the ImageOrientationPatient of the interested slice
  \item \code{orientationMatrix} is the complete orientationMatrix proposed as a matrix
  \item \code{StudyDate} the date of the study
  \item \code{Modality} the modality of the study
  \item \code{SeriesInstanceUID} the SeriesInstanceUID of the serie
}}

\item{seriesInstanceUID}{If the attribute is series-related, the seriesInstanceUID should be specified}

\item{fileName}{If the attribute is image-related, the fileName should be specified}
}
\value{
the whished attribute
}
\description{
It's a wrapper to get an attribute from a \code{geoLet} obj.
}
\examples{
\dontrun{
# create an object geoLet
obj<-geoLet()
# use the method 'openDICOMFolder' to load a study
obj$openDICOMFolder("/progetti/immagini/SaroshTest")
# use the wrapper GLT.getROIVoxels to get voxels of ROI "Ossa, NOS"
dataStorage<-GLT.getAttribute(obj = obj,attribute = "dataStorage")
# just to see the content of 'ossa'....
names(dataStorage)
}
}

