% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoLet.R
\name{GLT.getROIVoxels}
\alias{GLT.getROIVoxels}
\title{GLT.getROIVoxels - a wrapper function to get a set of voxel within a ROI from a \code{geoLet} object}
\usage{
GLT.getROIVoxels(obj = obj, Structure = Structure)
}
\arguments{
\item{obj}{an \code{geoLet} object}

\item{Structure}{the ROIName to extract}
}
\value{
a list
}
\description{
force a \code{geoLet} object to extract the CT/MR voxels within the interested ROI and return a voxelCube containing the CT/MR voxels, the mask and some info (as a list)
}
\details{
it's jusat a wrapper function to the \code{getROIVoxels} method of \code{geoLEt} class.
}
\examples{
\dontrun{
# create an object geoLet
obj<-geoLet()
# use the method 'openDICOMFolder' to load a study
obj$openDICOMFolder("/progetti/immagini/SaroshTest")
# use the wrapper GLT.getROIVoxels to get voxels of ROI "Ossa, NOS"
ossa<-GLT.getROIVoxels(obj = obj,Structure = "Ossa, NOS")
# just to see the content of 'ossa'....
names(ossa)
}
}

