% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoLet.R
\name{GLT.openDICOMFolder}
\alias{GLT.openDICOMFolder}
\title{GLT.openDICOMFolder - a wrapper function to force a geoLet object to load a DICOM Study}
\usage{
GLT.openDICOMFolder(obj = obj, pathToOpen = pathToOpen)
}
\arguments{
\item{obj}{an \code{geoLet} object}

\item{pathToOpen}{the path where the DICOM study is stored.}
}
\value{
nothing. To read the loaded data please retrieve the attribute \code{dataStorage} by the \code{getAttribute} method or by it's wrapper-function \code{GLT.getAttribute}.
}
\description{
force a \code{geoLet} object to load a DICOM study
}
\details{
it's jusat a wrapper function to force an mmButo object to load DICOM Studies by the method \code{openDICOMFolder}
}

