% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HI.generate}
\alias{HI.generate}
\title{Function that creates DVH matrix, 1st column is the dose bins, other columns are volume steps}
\usage{
HI.generate(dvhnumber, type = c("random", "convex", "concave", "mix"),
  dvh.type = c("differential", "cumulative"), vol.distr = c("relative",
  "absolute"), mindose = 0, maxdose = 75, dosebin = 0.5, mean.vol = 200)
}
\arguments{
\item{dvhnumber}{The number of DVHs to be generated}

\item{type}{The type of DVH to be generated: \code{"random"} generates all possible variation in dose distribution,
\code{"convex"} for DVH similar to PTV structures with good dose homogeneity to high levels,
\code{"concave"} for DVH with low level of dose delivered to the volume as a spared structure,
\code{"mix"} for DVH where dose distribution is averaged in the whole volume spectrum.}
}
\description{
Creates an object of class \code{dvhmatrix}.
}

