% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logHandler.R
\name{logHandler}
\alias{logHandler}
\title{class for handling logs/warnings/errors}
\usage{
logHandler()
}
\description{
It handles messages from script to a chosen output (screen, file, etc.)

              Many methods are available for objects built by this class:
              \itemize{
              \item \code{setOutput(attributeList)}
              sets the attributes of the object. Many attributes can be specified in the list:
              \itemize{
                \item \code{onFilePar} : \emph{logical}, default=\code{FALSE}. Set it to \code{TRUE} if you want the msgs  written on file;
                \item \code{onScreenPar} : \emph{logical}, default=\code{TRUE} Set it to \code{TRUE} if you want the msgs written on screen;
                \item \code{fileName} : \emph{string}, default='./defLogHandler.txt'. Set the file name where msgs should be written;
                \item \code{returnOnEOL} : \emph{logical}, default=\code{TRUE} Does it has to add a '\\n' at the end of the msg?
              }
              \item \code{sendLog(message)}
              print on screen, on file (or on a void output) the given message
              \item \code{do(what2Do, arguments="")}
              it send a command to the object. Many commands are available:
              \itemize{
                \item \code{clearOutputFile} : it truncates the output file. No arguments are needed.
              }
              }
}

