% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logHandlerRegister.R
\name{logHandlerRegister}
\alias{logHandlerRegister}
\title{class for handling a REGISTER for logs/warnings/errors}
\usage{
logHandlerRegister()
}
\description{
It provide a REGISTER for handling messages from script to a chosen output (screen, file, etc.)
}

