% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{mmButo}
\alias{mmButo}
\title{Load and handle a tree}
\usage{
mmButo()
}
\description{
this is the old (and deprecated) version of mmButo. It is able to look into a folder serching all DICOM studies and return a structure containint all the voxel cubes.
              The available methods are:
              \itemize{
              \item \code{openTreeMultiROIs(Path, structureList)}
              is a method used to open a chosen folder. This method loads all the DICOM objects into
              the indicated folder (without recursion) as attribute of the object.
              Information can be retrieved using \code{getAttribute} method or, simply, getting the result of this method.#'
              }
}

