% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmButo.R
\name{new.mmButo}
\alias{new.mmButo}
\title{Load and handle a tree}
\usage{
new.mmButo()
}
\description{
this is the old (and deprecated) version of mmButo. It is able to look into a folder serching all DICOM studies and return a structure containint all the voxel cubes.
              The available methods are:
}

