% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.areaVolume}
\alias{RAD.areaVolume}
\title{function to calculate Are/Volume and related measures}
\usage{
RAD.areaVolume(listaROIVoxels)
}
\arguments{
\item{listaROIVoxels}{an output of a \code{obj$getROIVoxel()} method}
}
\value{
a list containing, for each patient the indicated measures
}
\description{
calculates Are, Volume, Area/Volume Ratio and equivolumetric Spherical Area Ratio
}

