% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.compareSignals}
\alias{RAD.compareSignals}
\title{Apply erosion to a set of voxelCubes}
\usage{
RAD.compareSignals(ROIVoxelData, bootStrappedTimes = 1000)
}
\description{
Apply the erosion to a set of ROIs internal voxels
}

