% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.easyROI}
\alias{RAD.easyROI}
\title{return a simple list of voxelcubes}
\usage{
RAD.easyROI(ROIVoxelData)
}
\arguments{
\item{ROIVoxelData}{as got from of a \code{obj$getROIVoxel()} method. Cropped or expanded, it's the same}
}
\value{
a list containing the eroded voxelCubes presented in a 'easy to handle' shape
}
\description{
convertes the output of a \code{obj$getROIVoxel()} to an easy-to-handle list of voxelCubes
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
GTV<-obj$getROIVoxel(ROIName="GTV")

# get the possible biopsy
res<-RAD.easyROI( GTV )
}#'
}

