% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmButo.R
\name{RAD.execAlgorithm}
\alias{RAD.execAlgorithm}
\title{RAD.execAlgorithm - a wrapper function to force an mmButo object to execute calculations}
\usage{
RAD.execAlgorithm(obj, algorithm = c("KDF", "BAVA", "virtualBiopsy",
  "rawAreaVolume"), ROIName, grayTuningValue, ROIName4Tuning, nx = 2,
  ny = 2, nz = 0, errorHandlerParams = c())
}
\arguments{
\item{obj}{an \code{mmButo} object}

\item{algorithm}{a string indicating which algorithm should be run
\itemize{
  \item \code{KDF} is a Kernel Density Function in "Toronto style". This requires the specification of the \code{ROIName}, \code{grayTuningValue} and \code{ROIName4Tuning}
  \item \code{BAVA} it overlap all the histograms, resampling and syncing them to a same x-step and x-offest, after a common normalization to an upper bound value. This requires the specification of the \code{ROIName}, \code{grayTuningValue} and \code{ROIName4Tuning}.
  \item \code{virtualBiopsy} it does all the possible virtual Biopsies doable with a carrots of dimensions \code{nx},\code{ny},\code{nz} (if dimensions are not specified it uses \code{nx}=2, \code{ny}=2, \code{nz}=0 ). It provides also the computation of the mean and st.dev for all the carrots grouped for patient. This requires the specification of the \code{ROIName}, \code{grayTuningValue} and \code{ROIName4Tuning} and, optionale, \code{nx},\code{ny},\code{nz}
  \item \code{rawAreaVolume}
}}

\item{ROIName}{the name of the ROI you want to expose to calculus}

\item{grayTuningValue}{the "upper bound" for the normalization}

\item{ROIName4Tuning}{the name of the ROI you want to use for normalization}

\item{nx}{number of voxels along x axes}

\item{ny}{number of voxels along y axes}

\item{nz}{number of voxels along z axes}

\item{errorHandlerParams}{is a list to indicate what to do in case of error caught by moddicom (the error caught from R are not managed). The list can be so populated:
\itemize{
  \item \code{onFilePar} if set to \code{TRUE} the error/log messages will be written on a file;
  \item \code{fileName} if \code{onFilePar} is set to \code{TRUE} this attribute indicate the fileName. If it is not specificed the default is './defLogHandler.txt';
  \item \code{onScreenPar} if set to \code{TRUE} the error/log messages will be prompt on the screen;
  \item \code{returnOnEOL} if set to \code{TRUE} once the error is written, an End Of Line is added at the end of line.
}}
}
\value{
nothing. To read the loaded data please retrieve the attribute \code{results} by the \code{getAttribute} method or by it's wrapper-function \code{RAD.getAttribute}.
}
\description{
run the calculus for the given algorithm, for a given ROI, for all the cases stored in un mmButo object. Further runs of the same algorithm will override the results.
}
\details{
it's jusat a wrapper function of the method \code{execAlgorithm} of the \code{mmBute} class
}

