% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.getBiopsy}
\alias{RAD.getBiopsy}
\title{return the biopsy}
\usage{
RAD.getBiopsy(possBio, ROIVoxelData, x = 4, y = 4, z = 1)
}
\arguments{
\item{possBio}{as got from of a \code{RAD.VirtualBiopsy()} method.}

\item{ROIVoxelData}{is the voxelData (cropped) got from of a \code{obj$getROIVoxel()} method.}

\item{x}{the x dim of the biopsy (as passed to the \code{RAD.VirtualBiopsy()} method: the real cube will have (2*x+1) along x axes)}

\item{y}{the x dim of the biopsy (as passed to the \code{RAD.VirtualBiopsy()} method: the real cube will have (2*y+1) along y axes)}

\item{z}{the x dim of the biopsy (as passed to the \code{RAD.VirtualBiopsy()} method: the real cube will have (2*z+1) along z axes)}
}
\value{
a list (quite big) containing all the biopsy
}
\description{
get the list of possible centroids and return the list of biopsy
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
GTV<-obj$getROIVoxel(ROIName="GTV")

# get the possible centroids
a<-RAD.VirtualBiopsy(ROIVoxelData = GTV,dx.min = 3,dy.min = 4,dz.min = 1,dx.max = 4,dy.max = 4,dz.max = 1)

# get the biopsy
listaBiopsie<-RAD.getBiopsy(possBio = a,ROIVoxelData = GTV,x = 4,y = 4,z = 1)
}#'
}

