% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.imagesApplyFUN}
\alias{RAD.imagesApplyFUN}
\title{Calculate whatever you want from a defined region of mmButo voxel maps}
\usage{
RAD.imagesApplyFUN(obj.mmButo, ROIName, ROINameForNormalization = NA,
  collection = "default", biopsyDim.xyz = NA, erosion.xyz = NA,
  applyToBorder = FALSE, FUNFunction = NA, listOfFUNParameters = NA,
  cropResult = TRUE)
}
\arguments{
\item{obj.mmButo}{an object of class \code{new.mmButo}}

\item{ROIName}{the name of the interested ROI}

\item{collection}{the interested collection; the dafault collection is '\code{default}'}

\item{biopsyDim.xyz}{is a vector with the c(x,y,z) dimensions of the cube for biopsy. The default is \code{NA} which means that no biopsy has to be perfomed.}

\item{erosion.xyz}{is a vector with the c(x,y,z) dimensions of the erosion which should be applied to the voxelCubes respect the \code{ROIName.}}

\item{applyToBorder}{can be \code{TRUE} if you want to apply the FUNCTION to the border of \code{FALSE} if you want to apply the FUNCTION to all the voxel within the \code{ROIName} (eventually eroded)}

\item{FUNFunction}{is the FUNCTION which should be applied to any biopsy (or any single voxel)}

\item{listOfFUNParameters}{in order to avoid the 'pork party' due to the abominable use of '\code{...}', the parameters for the FUNCTION can be put in this \code{list}.}

\item{cropResult}{can be set to \code{TRUE} or \code{FALSE}. If \code{TRUE} the result will be cropped and will occupy much less RAM, otherwise it will have the original size (the default is \code{TRUE}).}

\item{normalizationROIName}{the name of the ROI used to normaliza the signal: default is \code{NA}}
}
\value{
a list containing the wished output
}
\description{
It applies a passed FUNCTION to the voxel maps of the new.mmButo objects. The maps can be the normal maps or maps obtained by erosion and/or biopsy
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

computeSDWithNoise<-function(a, listOfFUNParameters=list()) {
 b<-sd(a);
 if(listOfFUNParameters$noise==TRUE)  b<-b+b*runif(1)
 return(b)
}

# no noise, apply on an eroded GTV,
a<-RAD.imagesApplyFUN(obj.mmButo = obj,ROIName = "GTV",FUNFunction = calcolaSD,erosion.xyz = c(2,2,0),applyToBorder = FALSE,listOfFUNParameters = list("noise"=FALSE))

# no noise, apply on the BORDER
a<-RAD.imagesApplyFUN(obj.mmButo = obj,ROIName = "GTV",FUNFunction = calcolaSD,erosion.xyz = c(2,2,0),applyToBorder = TRUE,listOfFUNParameters = list("noise"=FALSE))

# no noise, not BORDER, on biopsy of <2,2,0> and erosion of <2,2,0>
a<-RAD.imagesApplyFUN(obj.mmButo = obj,ROIName = "GTV",FUNFunction = calcolaSD,erosion.xyz = c(5,5,3),applyToBorder = FALSE,listOfFUNParameters = list("noise"=FALSE),biopsyDim.xyz = c(2,2,0))

}#'
}

