% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.mmButo}
\alias{RAD.mmButo}
\title{Load and handle a tree}
\usage{
RAD.mmButo()
}
\description{
This function can be used to load, in one shot, many DICOM studies referring different patients. Each study is stored in a separate element of a list and basics analysis can be done. Even if it is quite deprecated, it can still be useful for 'quick & dirty' image analysis.
}

