% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RAD.MultiPIPOblique}
\alias{RAD.MultiPIPOblique}
\title{Calculates image voxels internal respect a given ROI}
\usage{
RAD.MultiPIPOblique(dataStorage, Structure, SeriesInstanceUID)
}
\arguments{
\item{dataStorage}{is the structure returned from a \code{obj$getAttribute("dataStructure")} where \code{obj} is an instance of a \code{geoLet}}

\item{Structure}{is the ROI name od the ROI that should be extracted}

\item{SeriesInstanceUID}{Is the interested series Instance UID.}
}
\value{
A list with unknown meaning
}
\description{
This function can be used to calculate internal voxels, for each ROIs, in a \code{dataStorage} data structure taken from a \code{geoLet} object
}

