% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.possibleVirtualBiopsy}
\alias{RAD.possibleVirtualBiopsy}
\title{Extract the possible Biopsies for a given ROIVoxelData}
\usage{
RAD.possibleVirtualBiopsy(ROIVoxelData, dx.min = 2, dy.min = 2,
  dz.min = 0, dx.max, dy.max, dz.max, sampleResultAt = Inf)
}
\arguments{
\item{ROIVoxelData}{as got from of a \code{obj$getROIVoxel()} method. It considers the cropped version and provide internally to explode it}

\item{dx.min}{the minumim number of voxels to be considered along the x dimension: default is equal to 2.}

\item{dy.min}{the minumim number of voxels to be considered along the y dimension: default is equal to 2}

\item{dz.min}{the minumim number of voxels to be considered along the z dimension: default is equal to 0'}

\item{dx.max}{the maximum number of voxel to be considered for biopsy along x axes;}

\item{dy.max}{the maximum number of voxel to be considered for biopsy along y axes;}

\item{dz.max}{the maximum number of voxel to be considered for biopsy along z axes;}

\item{sampleResultAt}{because of it is nurmaly not useful having the list of ALL the possible centroids (they can be hundred of thounsands) this parameters forces the function to RESAMPLE the list of possible centroids to the wished value. A value of 500 centroids, for example can be as espressive as a list of 100.000 centroids and save a lot of memory. Consider the adoption of this parameter if you don't want to be snowed under a tons of centoids (in particular for little volumes)}
}
\value{
a list containing a lot of things
}
\description{
Extract the possible Biopsies for a given ROIVoxelData. It can extract all the possible biopsies considering different volumes within a specified volume range along the axes. Please consider that if you give order to extract voxelCubes from dx.max=2,dy.max=2,dy.max=0 it will consider as biggest voxelCubes cube of 2*2+1,2*2+1,1 because with max.dx,max,dy,max,dz you specify only the distance from the centroid (the centroids itself is the '+1' in each direction)
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
Retto<-obj$getROIVoxel(ROIName="Retto")

# get the possible biopsy
biopsy<-RAD.VirtualBiopsy(dx.max = 3,dy.max = 3,dz.max = 2,ROIVoxelData = Retto,dx.min = 2,dy.min = 2, dz.min = 1, sampleResultAt = 500)
}#'
}

