% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmButo.R
\name{RAD.ROIStats}
\alias{RAD.ROIStats}
\title{RAD.ROIStats - a wrapper function for getting stats about stored ROIs}
\usage{
RAD.ROIStats(obj, ROIName, errorHandlerParams = c())
}
\arguments{
\item{obj}{an \code{mmButo} object}

\item{ROIName}{the name of the ROI you want to get stats}

\item{errorHandlerParams}{is a list to indicate what to do in case of error caught by moddicom (the error caught from R are not managed). The list can be so populated:
\itemize{
  \item \code{onFilePar} if set to \code{TRUE} the error/log messages will be written on a file;
  \item \code{fileName} if \code{onFilePar} is set to \code{TRUE} this attribute indicate the fileName. If it is not specificed the default is './defLogHandler.txt';
  \item \code{onScreenPar} if set to \code{TRUE} the error/log messages will be prompt on the screen;
  \item \code{returnOnEOL} if set to \code{TRUE} once the error is written, an End Of Line is added at the end of line.
}}
}
\value{
a list containing the details and a recap (total) of min/mean/max/stdDev of voxel within the given ROI
     \itemize{
       \item \code{detail-stdev} standard deviation of grey-voxel-values into the given ROI for the indicated Patient
       \item \code{detail-mean} mean of grey-voxel-values into the given ROI for the indicated Patient
       \item \code{detail-max} max of grey-voxel-values into the given ROI for the indicated Patient
       \item \code{total-minmean} the minimum from all the means from all the patients
       \item \code{total-meanmean} the mean of  all the means from all patients
       \item \code{total-maxmean} the max value from all the means from all the patients
       \item \code{total-stddevmean} the stdev value from all the means from all the patients
       \item \code{total-minmax} the min value from all the max from all the patients
       \item \code{total-meanmax} the mean value from all the max from all the patients
       \item \code{total-maxmax} the max value from all the max from all the patients
       \item \code{total-stdevmax} the stdev value from all the max from all the patients
     }
}
\description{
give back the stats of the chosen ROI
}

