% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmButo.R
\name{RAD.setAttribute}
\alias{RAD.setAttribute}
\title{RAD.openTreeMultiROIs - a wrapper function to force an mmButo object to load DICOM Studies}
\usage{
RAD.setAttribute(obj, attribute, value, errorHandlerParams = c())
}
\arguments{
\item{obj}{an \code{mmButo} object}

\item{attribute}{the name, as string, of the attribute you want to set
\itemize{
  \item \code{verbose} set the "verbose" level of the object
}}

\item{value}{the new value for the attribute}

\item{errorHandlerParams}{is a list to indicate what to do in case of error caught by moddicom (the error caught from R are not managed). The list can be so populated:
\itemize{
  \item \code{onFilePar} if set to \code{TRUE} the error/log messages will be written on a file;
  \item \code{fileName} if \code{onFilePar} is set to \code{TRUE} this attribute indicate the fileName. If it is not specificed the default is './defLogHandler.txt';
  \item \code{onScreenPar} if set to \code{TRUE} the error/log messages will be prompt on the screen;
  \item \code{returnOnEOL} if set to \code{TRUE} once the error is written, an End Of Line is added at the end of line.
}}
}
\value{
nothing. To read the loaded data please retrieve the attribute \code{dataStorage} by the \code{getAttribute} method or by it's wrapper-function \code{RAD.getAttribute}.
}
\description{
set an attribute of a \code{mmButo} object
}
\details{
it's jusat a wrapper function to force an mmButo object to load DICOM Studies by the method \code{openTreeMultiROIs}
}

