% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DVH.diff.to.cum}
\alias{DVH.diff.to.cum}
\title{Function that converts differential DVHs into cumulative ones}
\usage{
DVH.diff.to.cum(dvh, relative = TRUE)
}
\arguments{
\item{dvh}{Either an object of class \code{dvhmatrix} or \code{matrix} type.}

\item{relative}{If \code{TRUE} the structure volume bins are divided by the total volume.}
}
\value{
Either an object of class \code{dvhmatrix} or a \code{matrix} according the argument \code{dvh}.
}
\description{
Function that converts an object of class \code{dvhmatrix} or a simple \code{matrix} that
             represents a differential DVH into a cumulative one.
}

