% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Filter.R
\name{FIL.applyFilterToStudy}
\alias{FIL.applyFilterToStudy}
\title{Apply a filter to mmButo object}
\usage{
FIL.applyFilterToStudy(obj.mmButo, ROIVoxelData, filter.pipeline,
  collection = "default", cropResult = TRUE)
}
\arguments{
\item{obj.mmButo}{is the \code{new.mmButo} object}

\item{filter.pipeline}{is a \code{list} where is indicated the pipeline of filtering to be applied (in sequence)}

\item{collection}{is the collection: the default is '\code{default}'}

\item{cropResult}{is a boolean (\code{TRUE} or \code{FALSE}) which indicates if the result should be cropped or not, in order to save memory. Default is \code{TRUE}}

\item{ROINameForNormalization}{if specified indicates which ROI should be used for signal normalization}

\item{ROIName}{is the name of the ROI that should be extracted and filtered}
}
\value{
a list containing the filtered images (cropped or not)
}
\description{
It applies filter to an mButo object to simplify RADIOMICS issues
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection("/progetti/immagini/urinaEasy")

GTV<-obj$getROIVoxel(ROIName = "GTV")
Urina<-obj$getROIVoxel(ROIName = "Urina")
GTV.C<-obj$getCorrectedROIVoxel(inputROIVoxel = GTV,correctionROIVoxel = Urina,typeOfCorrection = "log")

# build the filtering pipeline
filterPipeline<- list()
filterPipeline[[1]]<-list("kernel.type"="gaussian")
filterPipeline[[2]]<-list("kernel.type"="emboss", "sigma"=1.5)

# filter the images (normalizing the GTV signal with Urin) cropping the result
a<-FIL.applyFilterToStudy(obj.mmButo = obj,ROIVoxelData = GTV.C, filter.pipeline = filterPipeline )

}#'
}

