% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoLet.R
\name{geoLet}
\alias{geoLet}
\title{class for loading and presenting DICOM data}
\usage{
geoLet()
}
\description{
Instantiate an object of the class \code{geoLet}.This represents just the classname,
              for each instantiated object many methods are available, not in canonical S3 or S4.

              The available methods are:
              \itemize{
              \item \code{openDICOMFolder(pathToOpen)}
              is a method used to open a chosen folder. This method loads all the DICOM objects into
              the indicated folder (without recursion) as attribute of the object.
              Information can be retrieved using \code{getAttribute} method or
              more specific methods.
              \item \code{getAttribute(attribute,seriesInstanceUID="",fileName="")}
              Is a method used to get internal attributes. Allowed values fot \code{attributes} are:

              \itemize{
                \item \code{dataStorage} : return a structured list with all the information retrieved from the
                DICOM object in the folder. More detailed information about the structur of the returned list
                are available at <inserire link>
                \item \code{ROIPointList} : return the ROI point List for all the stored ROIs;
                \item \code{PatientName} : return the content of the (0010,0010) DICOM tag(*);
                \item \code{PatientID} : return the content of the (0010,0020) DICOM tag(*);
                \item \code{Rows} : return the content of the (0028,0010) DICOM tag(*);
                \item \code{Columns} : return the content of the (0028,0011) DICOM tag(*);
                \item \code{StudyDate} : return the content of the (0028,0011) DICOM tag(*);
                \item \code{Modality} : return the content of the (0008,0060) DICOM tag(*);
                \item \code{PatientSex} : return the content of the (0010,0040) DICOM tag(*);
                \item \code{SeriesInstanceUID} : return the content of the (0020,000e) DICOM tag(*);
                \item \code{SliceThickness} : return the content of the (0018,0050) DICOM tag(*);
                \item \code{ImagePositionPatient} : return the content of the (0020,0032) DICOM tag(*);
                \item \code{ImageOrientationPatient} : return the content of the (0020,0037) DICOM tag(*);
                \item \code{PixelSpacing} : return the content of the (0028,0030) DICOM tag(*);
              }

              If no \code{seriesInstanceUID} or \code{filename} are provided, it returns the tag found in
              what seems to be the referencing CT or RMN scan. Pay attention to this point: in case of doubt about the
              content of the folder this can lead to errors.
              \item \code{getDICOMTag(fileName="",tag=tag) }
              it allows to retrieve the value of a specific DICOM tag on a specified DICOM filename. If the name is not
              indicated the method gets the first DICOM file he can find.
              \item \code{getROIList() }
              returns the list of the available ROIs
              \item \code{getROIPointList(ROINumber) }
              returns the list of the points which define the indicated ROI.
              \item \code{setAttribute(attributeName,attributeValue) }
              it sets the value of the specified attribute. The list of the availeable attributes is the following:
                 \itemize{
                   \item \code{verbose} : it define the policy adopted to print warnings and logs during computation. It is handled using an \code{errorHandler} object and works with three different levels. A list indicating the behaviour for each level should be provided.
                 }
              \item \code{getFolderContent(pathToOpen="") }
              explores the content of the given folder and returns informarion about the stored DICOM objects
              }
}

