\name{moddicom-package}
\alias{moddicom-package}
\alias{moddicom}
\docType{package}
\title{
DICOM RT files handler, dose/response models creation, ROI voxels analyzer
}
\description{
This package allows to load DICOM CT scans, MRIs, RTStruct, RTDose
and RTPlan. The aim of the package is to give a tool to perform
radiaotherapy dose analysis, handle voxels and ROI in a 3D space. A list of
wrappers for performing dose-response models analysis and fitting is given.
}
\details{
\tabular{ll}{
Package: \tab moddicom\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2014-12-01\cr
License: \tab GPL 2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Nicola Dinapoli, Roberto Gatta

Maintainer: Nicola Dinapoli <nicola.dinapoli@rm.unicatt.it>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
}
