% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.borderTextureMap}
\alias{RAD.borderTextureMap}
\title{Calculate entropy and stdev maps}
\usage{
RAD.borderTextureMap(obj.mmButo, ROIName, margin.x = 3, margin.y = 3,
  margin.z = 1, collection = "default", ROINameForNormalization = NA,
  erosion.x = 3, erosion.y = 3, erosion.z = 1, kindOfOutput = "normal")
}
\arguments{
\item{obj.mmButo}{an object of class \code{new.mmButo}}

\item{ROIName}{the name of the interested ROI}

\item{margin.x}{for the biopsy; default is 3 (the box has 7 voxels along x direction)}

\item{margin.y}{for the biopsy; default is 3 (the box has 7 voxels along y direction)}

\item{margin.z}{for the biopsy; default is 1 (the box has 3 voxels along z direction)}

\item{collection}{the interested collection; the dafault collection is '\code{default}'}

\item{erosion.x}{for the erosion; default is 3}

\item{erosion.y}{for the erosion; default is 3}

\item{erosion.z}{for the erosion; default is 1}

\item{kindOfOutput}{is a string which can be '\code{normal}' or '\code{extended}'. In the first case it returna a complex list, cropped, with all the information needed to extend the images; in the second it provides the extended images avoiding to return all the ancillary (geometry) data}

\item{normalizationROIName}{the name of the ROI used to normaliza the signal: default is \code{NA}}
}
\value{
a list containing the entropy maps and the stdev maps.
}
\description{
Calculate entropy maps and stdev maps of the images, eventually normalizing the image
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
Retto<-obj$getROIVoxel(ROIName="Retto")

# get the possible biopsy
erodedCubes<-RAD.applyErosion( ROIVoxelData = Retto )
}#'
}

