% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.firstOrderFeatureImage}
\alias{RAD.firstOrderFeatureImage}
\title{function to calculate the first order features}
\usage{
RAD.firstOrderFeatureImage(inputData, histSamples = 150)
}
\arguments{
\item{inputData}{a list where each element is an array of the voxel of the image. Each element of the list normally refers to a patient.}

\item{histSamples}{in order to avoid distorsion in the calculus of the entropy the histogram can be sampled to a wished number of bins. The defaul is 150, any other integer value can be set using this parameter.}
}
\value{
three lists: the first list contains the entropies, the second the kurtosis and the third the skewness
}
\description{
calculates Shannon entropy, kursosis and Skewness of a given list of arrays of voxels
}

