% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.firstOrderFeatureImage}
\alias{RAD.firstOrderFeatureImage}
\title{function to calculate the first order features}
\usage{
RAD.firstOrderFeatureImage(inputData)
}
\arguments{
\item{inputData}{a list where each element is an array of the voxel of the image. Each element of the list normally refers to a patient.}
}
\value{
three lists: the first list contains the entropies, the second the kurtosis and the third the skewness
}
\description{
calculates Shannon entropy, kursosis and Skewness of a given list of arrays of voxels
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
Retto<-obj$getROIVoxel(ROIName="Retto")

# get the possible biopsy
aa<-RAD.firstOrderFeatureImage(inputData = Retto )
aa$entropy
}#' #'
}

