% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radiomics.R
\name{RAD.imagesApplyFUN}
\alias{RAD.imagesApplyFUN}
\title{Apply a user defined function over a defined region of mmButo voxel maps}
\usage{
RAD.imagesApplyFUN(obj.mmButo, ROIVoxelData, collection = "default",
  biopsyDim.xyz = NA, erosion.xyz = NA, applyToBorder = FALSE, FUN = NA,
  cropResult = TRUE, ...)
}
\arguments{
\item{obj.mmButo}{an object of class \code{new.mmButo}. Ignored if \code{ROIName} and \code{normalizationROIName} are not character but \code{new.mmButo$getROIVoxel()} object types}

\item{ROIVoxelData}{the ROIVoxelData as obtained from a \code{getROIVoxel()} or a \code{getCorrectedROIVoxel}}

\item{collection}{the interested collection; the dafault collection is '\code{default}'}

\item{biopsyDim.xyz}{is an integer numeric vector with the \code{c(x,y,z)} dimensions in voxels of the cube for biopsy. The default is \code{NA} which means that no biopsy is perfomed}

\item{erosion.xyz}{is an integer vector with the \code{c(x,y,z)} dimensions of the erosion which has to be applied to the voxelCubes respect the \code{ROIName}}

\item{applyToBorder}{a logical value, if \code{TRUE} the function \code{FUN} is applied to the border, if \code{FALSE} the FUNCTION is applied to all the voxel within the \code{ROIName} (eventually eroded)}

\item{FUN}{is the \code{function} to be be applied to any biopsy (if a \code{biopsyDim.xyz}) is given, otherwise the function is applied to all voxels}

\item{cropResult}{a logical value. If \code{TRUE} the result will be cropped and will occupy much less RAM, otherwise it will have the original size (default value is \code{TRUE})}

\item{...}{Other parameters to be passed to the function \code{FUN}}
}
\value{
a list containing the output of the function \code{FUN} applied over biopsies or over the whole ROI voxels series#'
}
\description{
It applies a user defined \code{function} over a voxel maps of in a \code{new.mmButo} object. The maps can be the normal maps or maps obtained by erosion and/or biopsy.
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
obj<-new.mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

Urina<-obj$getROIVoxel(ROIName = "Urina")
GTV<-obj$getROIVoxel(ROIName = "GTV")
GTV.C <- obj$getCorrectedROIVoxel(inputROIVoxel = GTV,correctionROIVoxel = Urina,typeOfCorrection = "log")

calcolaSD<-function(a, listOfFUNParameters=list()) {
 b<-sd(a);
 if(listOfFUNParameters$noise==TRUE)  b<-b+b*runif(1)
 return(b)
}

# no noise, not BORDER, on biopsy of <2,2,0> and erosion of <2,2,0>
a<-RAD.imagesApplyFUN(obj.mmButo = obj,ROIVoxelData = GTV.C,FUN = calcolaSD,erosion.xyz = c(5,5,3),applyToBorder = FALSE,listOfFUNParameters = list("noise"=FALSE),biopsyDim.xyz = c(2,2,0))

}#'
}

